Samlet alle tester og laget egne for BAKS
