+++ src/frontend/pages/fagsak/vilkaarsvurdering/gammel-vilkårsvurdering/vilkaarsvurdering-periode/VilkårsvurderingPeriodeSkjema.test.tsx
@@ -1,15 +1,18 @@
 import type { VilkårsvurderingPeriodeSkjemaData } from '../typer/vilkårsvurdering';
 import type { RenderResult } from '@testing-library/react';
 import type { UserEvent } from '@testing-library/user-event';
+import type { BehandlingDto, FagsakDto } from '~/generated';
 import type { VilkårsvurderingHook } from '~/pages/fagsak/vilkaarsvurdering/gammel-vilkårsvurdering/VilkårsvurderingContext';
 
 import { QueryClientProvider } from '@tanstack/react-query';
-import { render, waitFor } from '@testing-library/react';
+import { render, waitFor, within } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
+import { vi } from 'vitest';
 
 import { FagsakContext } from '~/context/FagsakContext';
 import { Aktsomhet, SærligeGrunner, Vilkårsresultat } from '~/kodeverk';
 import { TestBehandlingProvider } from '~/testdata/behandlingContextFactory';
+import { lagBehandling } from '~/testdata/behandlingFactory';
 import { lagFagsak } from '~/testdata/fagsakFactory';
 import { lagVilkårsvurderingPeriodeSkjemaData } from '~/testdata/vilkårsvurderingFactory';
 import { createTestQueryClient } from '~/testutils/queryTestUtils';
@@ -17,11 +20,12 @@ import { SkalUnnlates } from '~/typer/tilbakekrevingstyper';
 
 import { VilkårsvurderingPeriodeSkjema } from './VilkårsvurderingPeriodeSkjema';
 
-vi.setConfig({ testTimeout: 10000 });
+const mockKanIlleggeRenter = vi.hoisted(() => ({ value: true }));
 
 vi.mock('../VilkårsvurderingContext', () => {
     return {
         useVilkårsvurdering: (): Partial<VilkårsvurderingHook> => ({
+            kanIlleggeRenter: mockKanIlleggeRenter.value,
             oppdaterPeriode: vi.fn(),
             navigerTilNeste: vi.fn(),
             sendInnSkjemaOgNaviger: vi.fn(),
@@ -35,21 +39,29 @@ vi.mock('../VilkårsvurderingContext', () => {
     };
 });
 
-const renderVilkårsvurderingPeriodeSkjema = (
-    periode: VilkårsvurderingPeriodeSkjemaData,
-    erTotalbeløpUnder4Rettsgebyr: boolean,
-    behandletPerioder: VilkårsvurderingPeriodeSkjemaData[] = []
-): RenderResult => {
+const renderVilkårsvurderingPeriodeSkjema = ({
+    periode,
+    erTotalbeløpUnder4Rettsgebyr,
+    behandletPerioder,
+    behandling,
+    fagsak,
+}: {
+    periode?: VilkårsvurderingPeriodeSkjemaData;
+    erTotalbeløpUnder4Rettsgebyr?: boolean;
+    behandletPerioder?: VilkårsvurderingPeriodeSkjemaData[];
+    behandling?: BehandlingDto;
+    fagsak?: FagsakDto;
+}): RenderResult => {
     const queryClient = createTestQueryClient();
     return render(
         <QueryClientProvider client={queryClient}>
-            <FagsakContext value={lagFagsak()}>
-                <TestBehandlingProvider>
+            <FagsakContext value={fagsak ?? lagFagsak()}>
+                <TestBehandlingProvider behandling={behandling}>
                     <VilkårsvurderingPeriodeSkjema
-                        periode={periode}
-                        behandletPerioder={behandletPerioder}
-                        erTotalbeløpUnder4Rettsgebyr={erTotalbeløpUnder4Rettsgebyr}
-                        perioder={[periode]}
+                        periode={periode ?? lagVilkårsvurderingPeriodeSkjemaData()}
+                        behandletPerioder={behandletPerioder ?? []}
+                        erTotalbeløpUnder4Rettsgebyr={erTotalbeløpUnder4Rettsgebyr ?? false}
+                        perioder={[periode ?? lagVilkårsvurderingPeriodeSkjemaData()]}
                         pendingPeriode={undefined}
                         settPendingPeriode={vi.fn()}
                     />
@@ -64,6 +76,7 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     beforeEach(() => {
         user = userEvent.setup();
         vi.clearAllMocks();
+        mockKanIlleggeRenter.value = true;
     });
 
     test('God tro - beløp ikke i behold', async () => {
@@ -79,17 +92,22 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
                 },
             ],
         });
-        const behandletPerioder = [lagVilkårsvurderingPeriodeSkjemaData()];
-
         const { getByLabelText, getByRole, getByText, queryAllByText, queryByLabelText } =
-            renderVilkårsvurderingPeriodeSkjema(vilkårsvurderingPeriode, false, behandletPerioder);
+            renderVilkårsvurderingPeriodeSkjema({
+                periode: vilkårsvurderingPeriode,
+                erTotalbeløpUnder4Rettsgebyr: false,
+                behandletPerioder: [lagVilkårsvurderingPeriodeSkjemaData()],
+            });
 
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
         expect(getByText('Feilutbetalt beløp')).toBeInTheDocument();
         expect(getByText('Aktivitet 1')).toBeInTheDocument();
         expect(getByText('1 333')).toBeInTheDocument();
         expect(getByText('Aktivitet 2')).toBeInTheDocument();
         expect(getByText('1 000')).toBeInTheDocument();
+
+        expect(
+            queryByLabelText('Begrunn hvorfor du valgte alternativet ovenfor')
+        ).not.toBeInTheDocument();
 
         expect(
             getByRole('radio', {
@@ -127,11 +145,15 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
             })
         ).toBeEnabled();
 
-        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
-
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
             'begrunnelse'
+        );
+        await user.click(
+            getByLabelText('Mottaker har mottatt beløpet i aktsom god tro', {
+                selector: 'input',
+                exact: false,
+            })
         );
 
         expect(
@@ -139,21 +161,6 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
                 name: 'Gå videre til vedtakssteget',
             })
         ).toBeEnabled();
-
-        await user.click(
-            getByRole('button', {
-                name: 'Gå videre til vedtakssteget',
-            })
-        );
-
-        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(1);
-
-        await user.click(
-            getByLabelText('Mottaker har mottatt beløpet i aktsom god tro', {
-                selector: 'input',
-                exact: false,
-            })
-        );
 
         expect(
             queryByLabelText('Begrunn hvorfor beløpet ikke er i behold')
@@ -186,17 +193,14 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     });
 
     test('God tro - beløp i behold', async () => {
-        const {
-            getByLabelText,
-            getByRole,
-            getByText,
-            queryAllByText,
-            queryByLabelText,
-            queryByText,
-        } = renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
+        const { getByLabelText, getByRole, queryAllByText, queryByLabelText, queryByText } =
+            renderVilkårsvurderingPeriodeSkjema({});
 
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
         expect(queryByText('Kopier vilkårsvurdering fra')).not.toBeInTheDocument();
+
+        expect(
+            queryByLabelText('Begrunn hvorfor du valgte alternativet ovenfor')
+        ).not.toBeInTheDocument();
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -243,9 +247,8 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
             queryAllByText,
             queryByLabelText,
             queryByText,
-        } = renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
+        } = renderVilkårsvurderingPeriodeSkjema({});
 
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
         expect(
             queryByLabelText('Begrunn hvorfor du valgte alternativet ovenfor')
         ).not.toBeInTheDocument();
@@ -290,29 +293,29 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         expect(getByText('Andel som skal tilbakekreves')).toBeInTheDocument();
         expect(getByText('100%')).toBeInTheDocument();
         expect(getByText('Skal det beregnes 10% rentetillegg?')).toBeInTheDocument();
-        expect(getByText('Nei')).toBeInTheDocument();
+        expect(getByLabelText('Nei')).toBeChecked();
 
         await user.click(
             getByRole('button', {
                 name: 'Gå videre til vedtakssteget',
             })
         );
+
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
     });
 
-    test('Forsto/burde forstått - Mottaker må ha forstått at utbetalingen skyldtes en feil - ingen grunn til reduksjon', async () => {
+    test('Forsto/burde forstått - må ha forstått - ingen grunn til reduksjon', async () => {
         const {
             getByLabelText,
             getByRole,
+            getByTestId,
             getByText,
             queryAllByText,
             queryByLabelText,
             queryByRole,
             queryByText,
-            getByTestId,
-        } = renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
+        } = renderVilkårsvurderingPeriodeSkjema({});
 
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
         expect(
             queryByLabelText('Begrunn hvorfor du valgte alternativet ovenfor')
         ).not.toBeInTheDocument();
@@ -400,7 +403,6 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         );
 
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(1);
-        expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(0);
 
         await user.type(
             getByRole('textbox', {
@@ -418,10 +420,8 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     });
 
     test('Feilaktig - forsto', async () => {
-        const { getByLabelText, getByRole, getByText, queryAllByText, getByTestId } =
-            renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+        const { getByLabelText, getByRole, getByText, queryAllByText, queryByText, getByTestId } =
+            renderVilkårsvurderingPeriodeSkjema({});
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -453,7 +453,55 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
 
         expect(getByText('Andel som skal tilbakekreves')).toBeInTheDocument();
         expect(getByText('100%')).toBeInTheDocument();
-        expect(getByText('Skal det beregnes 10% rentetillegg?')).toBeInTheDocument();
+        expect(queryByText('Skal det beregnes 10% rentetillegg?')).toBeInTheDocument();
+        expect(getByTestId('skalDetTilleggesRenter_Ja')).toBeChecked();
+
+        await user.click(
+            getByRole('button', {
+                name: 'Gå videre til vedtakssteget',
+            })
+        );
+        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
+    });
+
+    test('BAKS - Feilaktig - forsto', async () => {
+        mockKanIlleggeRenter.value = false; // Mocker context som leser ytelsestype enten Barnetrygd eller Kontantstøtte fra useFagsak
+        const { getByLabelText, getByRole, getByText, queryAllByText, queryByText, getByTestId } =
+            renderVilkårsvurderingPeriodeSkjema({
+                fagsak: lagFagsak({ fagsystem: 'BA', ytelsestype: 'BARNETRYGD' }),
+            });
+
+        await user.type(
+            getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
+            'begrunnelse'
+        );
+        await user.click(
+            getByLabelText(
+                'Mottaker har forårsaket feilutbetalingen ved forsett eller uaktsomt gitt feilaktige opplysninger',
+                {
+                    selector: 'input',
+                    exact: false,
+                }
+            )
+        );
+
+        await user.click(
+            getByRole('button', {
+                name: 'Gå videre til vedtakssteget',
+            })
+        );
+        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(2);
+
+        await user.type(getByLabelText('Begrunn mottakerens aktsomhetsgrad'), 'begrunnelse');
+        await user.click(
+            getByLabelText('Forsettlig', {
+                selector: 'input',
+            })
+        );
+
+        expect(getByText('Andel som skal tilbakekreves')).toBeInTheDocument();
+        expect(getByText('100%')).toBeInTheDocument();
+        expect(queryByText('Skal det beregnes 10% rentetillegg?')).toBeInTheDocument();
         expect(getByTestId('skalDetTilleggesRenter_Nei')).toBeChecked();
 
         await user.click(
@@ -465,10 +513,8 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     });
 
     test('Feilaktige - grovt uaktsomt - ingen grunn til reduksjon', async () => {
-        const { getByLabelText, getByRole, getByText, queryAllByText, getByTestId } =
-            renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+        const { getByLabelText, getByRole, getByText, queryAllByText } =
+            renderVilkårsvurderingPeriodeSkjema({});
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -512,31 +558,25 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         );
 
         expect(getByText('Andel som skal tilbakekreves')).toBeInTheDocument();
-        expect(getByText('Skal det beregnes 10% rentetillegg?')).toBeInTheDocument();
-        expect(getByTestId('skalDetTilleggesRenter_Nei')).toBeChecked();
+        const rentetilleggRadioGroup = getByRole('radiogroup', {
+            name: 'Skal det beregnes 10% rentetillegg?',
+        });
+        expect(within(rentetilleggRadioGroup).getByRole('radio', { name: 'Ja' })).toBeChecked();
 
         await user.click(
             getByRole('button', {
                 name: 'Gå videre til vedtakssteget',
             })
         );
-        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
+
         expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(0);
+        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
     });
 
-    test('Feilaktige - grovt uaktsomt - grunn til reduksjon', async () => {
-        const {
-            getByLabelText,
-            getByRole,
-            getByText,
-            queryAllByText,
-            queryByLabelText,
-            queryByText,
-            getByTestId,
-        } = renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
-        expect(queryByText('Vurder mottakers grad av aktsomhet')).not.toBeInTheDocument();
+    test('BAKS - Feilaktige - grovt uaktsomt - ingen grunn til reduksjon', async () => {
+        mockKanIlleggeRenter.value = false; // Mocker context som leser ytelsestype enten Barnetrygd eller Kontantstøtte fra useFagsak
+        const { getByLabelText, getByRole, getByText, queryAllByText } =
+            renderVilkårsvurderingPeriodeSkjema({});
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -552,26 +592,12 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
             )
         );
 
-        expect(
-            queryByLabelText('Begrunn resultatet av vurderingen ovenfor')
-        ).not.toBeInTheDocument();
-        expect(
-            queryByText('Hvilke særlige grunner kan være aktuelle i denne saken?')
-        ).not.toBeInTheDocument();
-        expect(queryByText('Skal særlige grunner redusere beløpet?')).not.toBeInTheDocument();
-
         await user.type(getByLabelText('Begrunn mottakerens aktsomhetsgrad'), 'begrunnelse');
         await user.click(
             getByLabelText('Grovt uaktsomt', {
                 selector: 'input',
             })
         );
-
-        expect(queryByLabelText('Begrunn resultatet av vurderingen ovenfor')).toBeInTheDocument();
-        expect(
-            queryByText('Hvilke særlige grunner kan være aktuelle i denne saken?')
-        ).toBeInTheDocument();
-        expect(queryByText('Skal særlige grunner redusere beløpet?')).toBeInTheDocument();
 
         await user.click(
             getByRole('button', {
@@ -589,18 +615,90 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         );
         await user.click(
             getByRole('radio', {
-                name: 'Ja',
+                name: 'Nei',
             })
         );
 
-        expect(getByText('Angi andel som skal tilbakekreves')).toBeInTheDocument();
-        expect(getByText('Skal det beregnes 10% rentetillegg?')).toBeInTheDocument();
-        expect(getByTestId('skalDetTilleggesRenter_Nei')).toBeChecked();
-        expect(
-            getByRole('combobox', {
-                name: 'Angi andel som skal tilbakekreves',
+        expect(getByText('Andel som skal tilbakekreves')).toBeInTheDocument();
+        const rentetilleggRadioGroup = getByRole('radiogroup', {
+            name: 'SkrivebeskyttetSkal det beregnes 10% rentetillegg?',
+        });
+        expect(within(rentetilleggRadioGroup).getByRole('radio', { name: 'Nei' })).toBeChecked();
+
+        await user.click(
+            getByRole('button', {
+                name: 'Gå videre til vedtakssteget',
             })
-        ).toBeInTheDocument();
+        );
+
+        expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(0);
+        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
+    });
+
+    test('Feilaktige - grovt uaktsomt - grunn til reduksjon', async () => {
+        const { getByLabelText, getByRole, queryByLabelText, queryAllByText, queryByText } =
+            renderVilkårsvurderingPeriodeSkjema({});
+
+        expect(
+            queryByLabelText('Begrunn hvorfor du valgte alternativet ovenfor')
+        ).not.toBeInTheDocument();
+        expect(
+            queryByText('I hvilken grad har mottaker handlet uaktsomt?')
+        ).not.toBeInTheDocument();
+
+        const vilkårRadioGroup = getByRole('radiogroup', {
+            name: 'Velg det vilkåret i folketrygdloven §22-15 som gjelder for perioden',
+        });
+        await user.click(
+            within(vilkårRadioGroup).getByRole('radio', {
+                name: /Mottaker har forårsaket feilutbetalingen ved forsett eller uaktsomt gitt feilaktige opplysninger/i,
+            })
+        );
+        await user.type(
+            getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
+            'begrunnelse'
+        );
+
+        expect(
+            queryByLabelText('Begrunn resultatet av vurderingen ovenfor')
+        ).not.toBeInTheDocument();
+        expect(
+            queryByText('Hvilke særlige grunner kan være aktuelle i denne saken?')
+        ).not.toBeInTheDocument();
+        expect(queryByText('Skal særlige grunner redusere beløpet?')).not.toBeInTheDocument();
+
+        const aktsomhetsgradRadioGroup = getByRole('radiogroup', {
+            name: 'I hvilken grad har mottaker handlet uaktsomt?',
+        });
+        await user.click(
+            within(aktsomhetsgradRadioGroup).getByRole('radio', {
+                name: 'Grovt uaktsomt',
+            })
+        );
+        await user.type(getByLabelText('Begrunn mottakerens aktsomhetsgrad'), 'begrunnelse');
+
+        await user.click(
+            getByRole('button', {
+                name: 'Gå videre til vedtakssteget',
+            })
+        );
+        expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(1);
+        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(2);
+
+        const særligeGrunnerCheckboxGroup = getByRole('group', {
+            name: 'Hvilke særlige grunner kan være aktuelle i denne saken?',
+        });
+        await user.click(
+            within(særligeGrunnerCheckboxGroup).getByRole('checkbox', {
+                name: 'Graden av uaktsomhet hos den som kravet retter seg mot',
+            })
+        );
+
+        const særligeGrunnerRadioGroup = getByRole('radiogroup', {
+            name: 'Skal særlige grunner redusere beløpet?',
+        });
+        await user.click(within(særligeGrunnerRadioGroup).getByRole('radio', { name: 'Ja' }));
+        await user.type(getByLabelText('Begrunn resultatet av vurderingen ovenfor'), 'begrunnelse');
 
         await user.click(
             getByRole('button', {
@@ -608,14 +706,16 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
             })
         );
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(1);
-        expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(0);
 
-        await user.selectOptions(
-            getByRole('combobox', {
-                name: 'Angi andel som skal tilbakekreves',
-            }),
-            '30'
-        );
+        const andelAvBeløp = getByRole('combobox', {
+            name: 'Angi andel som skal tilbakekreves',
+        });
+        await user.selectOptions(andelAvBeløp, '30');
+
+        const rentetilleggRadioGroup = getByRole('radiogroup', {
+            name: 'SkrivebeskyttetSkal det beregnes 10% rentetillegg?',
+        });
+        expect(within(rentetilleggRadioGroup).getByRole('radio', { name: 'Ja' })).toBeChecked();
 
         await user.click(
             getByRole('button', {
@@ -625,70 +725,148 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
     });
 
+    test('nyModell - Feilaktige - grovt uaktsomt - låst ja rentetillegg', async () => {
+        const { getByRole } = renderVilkårsvurderingPeriodeSkjema({
+            behandling: lagBehandling({ erNyModell: true }),
+        });
+
+        const vilkårRadioGroup = getByRole('radiogroup', {
+            name: 'Velg det vilkåret i folketrygdloven §22-15 som gjelder for perioden',
+        });
+        await user.click(
+            within(vilkårRadioGroup).getByRole('radio', {
+                name: /Mottaker har forårsaket feilutbetalingen ved forsett eller uaktsomt gitt feilaktige opplysninger/i,
+            })
+        );
+
+        const aktsomhetsgradRadioGroup = getByRole('radiogroup', {
+            name: 'I hvilken grad har mottaker handlet uaktsomt?',
+        });
+        await user.click(
+            within(aktsomhetsgradRadioGroup).getByRole('radio', {
+                name: 'Grovt uaktsomt',
+            })
+        );
+        const særligeGrunnerRadioGroup = getByRole('radiogroup', {
+            name: 'Skal særlige grunner redusere beløpet?',
+        });
+        await user.click(within(særligeGrunnerRadioGroup).getByRole('radio', { name: 'Ja' }));
+
+        const rentetilleggRadioGroup = getByRole('radiogroup', {
+            name: 'SkrivebeskyttetSkal det beregnes 10% rentetillegg?',
+        });
+        expect(within(rentetilleggRadioGroup).getByRole('radio', { name: 'Ja' })).toBeChecked();
+    });
+
+    test('nyModell - ForstoBurdeForstått - forsto - ikke synlig rentetillegg', async () => {
+        const { getByRole, queryByText } = renderVilkårsvurderingPeriodeSkjema({
+            behandling: lagBehandling({ erNyModell: true }),
+        });
+
+        const vilkårRadioGroup = getByRole('radiogroup', {
+            name: 'Velg det vilkåret i folketrygdloven §22-15 som gjelder for perioden',
+        });
+        await user.click(
+            within(vilkårRadioGroup).getByRole('radio', {
+                name: 'Mottaker forsto eller burde forstått at utbetalingen skyldtes en feil (1. ledd, 1. punktum)',
+            })
+        );
+        const aktsomhetsgradRadioGroup = getByRole('radiogroup', {
+            name: 'Vurder mottakers grad av aktsomhet',
+        });
+        await user.click(
+            within(aktsomhetsgradRadioGroup).getByRole('radio', {
+                name: 'Mottaker forsto at utbetalingen skyldtes en feil',
+            })
+        );
+
+        expect(queryByText('Skal det beregnes 10% rentetillegg?')).not.toBeInTheDocument();
+    });
+
+    test('nyModell - ForstoBurdeForstått - må ha forstått - ikke synlig rentetillegg', async () => {
+        const { getByRole, queryByText } = renderVilkårsvurderingPeriodeSkjema({
+            behandling: lagBehandling({ erNyModell: true }),
+        });
+
+        const vilkårRadioGroup = getByRole('radiogroup', {
+            name: 'Velg det vilkåret i folketrygdloven §22-15 som gjelder for perioden',
+        });
+        await user.click(
+            within(vilkårRadioGroup).getByRole('radio', {
+                name: 'Mottaker forsto eller burde forstått at utbetalingen skyldtes en feil (1. ledd, 1. punktum)',
+            })
+        );
+        const aktsomhetsgradRadioGroup = getByRole('radiogroup', {
+            name: 'Vurder mottakers grad av aktsomhet',
+        });
+        await user.click(
+            within(aktsomhetsgradRadioGroup).getByRole('radio', {
+                name: 'Mottaker må ha forstått at utbetalingen skyldtes en feil',
+            })
+        );
+
+        const særligeGrunnerRadioGroup = getByRole('radiogroup', {
+            name: 'Skal særlige grunner redusere beløpet?',
+        });
+        await user.click(within(særligeGrunnerRadioGroup).getByRole('radio', { name: 'Ja' }));
+
+        expect(queryByText('Skal det beregnes 10% rentetillegg?')).not.toBeInTheDocument();
+    });
+
     test('Feilaktige - grovt uaktsomt - grunn til reduksjon - egendefinert', async () => {
-        const { getByLabelText, getByRole, getByText, queryAllByText, queryByRole } =
-            renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), false);
+        const { getByLabelText, getByRole, queryAllByText } = renderVilkårsvurderingPeriodeSkjema(
+            {}
+        );
 
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
-
+        const vilkårRadioGroup = getByRole('radiogroup', {
+            name: 'Velg det vilkåret i folketrygdloven §22-15 som gjelder for perioden',
+        });
+        await user.click(
+            within(vilkårRadioGroup).getByRole('radio', {
+                name: /Mottaker har forårsaket feilutbetalingen ved forsett eller uaktsomt gitt feilaktige opplysninger/i,
+            })
+        );
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
             'begrunnelse'
         );
-        await user.click(
-            getByLabelText(
-                'Mottaker har forårsaket feilutbetalingen ved forsett eller uaktsomt gitt feilaktige opplysninger',
-                {
-                    selector: 'input',
-                    exact: false,
-                }
-            )
-        );
 
+        const aktsomhetsgradRadioGroup = getByRole('radiogroup', {
+            name: 'I hvilken grad har mottaker handlet uaktsomt?',
+        });
+        await user.click(
+            within(aktsomhetsgradRadioGroup).getByRole('radio', {
+                name: 'Grovt uaktsomt',
+            })
+        );
         await user.type(getByLabelText('Begrunn mottakerens aktsomhetsgrad'), 'begrunnelse');
+
+        const særligeGrunnerCheckboxGroup = getByRole('group', {
+            name: 'Hvilke særlige grunner kan være aktuelle i denne saken?',
+        });
         await user.click(
-            getByLabelText('Grovt uaktsomt', {
-                selector: 'input',
+            within(særligeGrunnerCheckboxGroup).getByRole('checkbox', {
+                name: 'Graden av uaktsomhet hos den som kravet retter seg mot',
             })
         );
 
+        const særligeGrunnerRadioGroup = getByRole('radiogroup', {
+            name: 'Skal særlige grunner redusere beløpet?',
+        });
+        await user.click(within(særligeGrunnerRadioGroup).getByRole('radio', { name: 'Ja' }));
         await user.type(getByLabelText('Begrunn resultatet av vurderingen ovenfor'), 'begrunnelse');
-        await user.click(
-            getByLabelText('Graden av uaktsomhet hos den som kravet retter seg mot', {
-                selector: 'input',
-            })
-        );
-        await user.click(
-            getByRole('radio', {
-                name: 'Ja',
-            })
-        );
 
-        expect(
-            getByRole('combobox', {
-                name: 'Angi andel som skal tilbakekreves',
-            })
-        ).toBeInTheDocument();
-
+        const andelAvBeløp = getByRole('combobox', {
+            name: 'Angi andel som skal tilbakekreves',
+        });
         await user.click(
             getByRole('button', {
                 name: 'Gå videre til vedtakssteget',
             })
         );
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(1);
-        expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(0);
-        expect(
-            queryByRole('textbox', {
-                name: '',
-            })
-        ).not.toBeInTheDocument();
 
-        await user.selectOptions(
-            getByRole('combobox', {
-                name: 'Angi andel som skal tilbakekreves',
-            }),
-            'Egendefinert'
-        );
+        await user.selectOptions(andelAvBeløp, 'Egendefinert');
 
         await user.click(
             getByRole('button', {
@@ -697,12 +875,10 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         );
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(1);
 
-        await user.type(
-            getByRole('textbox', {
-                name: 'Angi andel som skal tilbakekreves - fritekst',
-            }),
-            '22'
-        );
+        const andelAvBeløpFritekst = getByRole('textbox', {
+            name: 'Angi andel som skal tilbakekreves - fritekst',
+        });
+        await user.type(andelAvBeløpFritekst, '22');
 
         await user.click(
             getByRole('button', {
@@ -717,13 +893,10 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
             getByLabelText,
             getByRole,
             getByTestId,
-            getByText,
             queryAllByText,
             queryByLabelText,
             queryByText,
-        } = renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), true);
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+        } = renderVilkårsvurderingPeriodeSkjema({ erTotalbeløpUnder4Rettsgebyr: true });
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -806,7 +979,6 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
             })
         );
         expect(queryAllByText('Feltet må fylles ut')).toHaveLength(1);
-        expect(queryAllByText('Du må velge minst en særlig grunn')).toHaveLength(0);
 
         await user.selectOptions(
             getByRole('combobox', {
@@ -825,9 +997,7 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
 
     test('Mangelfulle - uaktsomt - under 4 rettsgebyr - ingen grunn til reduksjon', async () => {
         const { getByLabelText, getByRole, getByText, getByTestId, queryAllByText, queryByText } =
-            renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), true);
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+            renderVilkårsvurderingPeriodeSkjema({ erTotalbeløpUnder4Rettsgebyr: true });
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -905,10 +1075,8 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     });
 
     test('Mangelfulle - uaktsomt - under 4 rettsgebyr - ikke tilbakekreves', async () => {
-        const { getByLabelText, getByRole, getByText, queryAllByText, queryByText } =
-            renderVilkårsvurderingPeriodeSkjema(lagVilkårsvurderingPeriodeSkjemaData(), true);
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+        const { getByLabelText, getByRole, queryAllByText, queryByText } =
+            renderVilkårsvurderingPeriodeSkjema({ erTotalbeløpUnder4Rettsgebyr: true });
 
         await user.type(
             getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor'),
@@ -966,22 +1134,22 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     });
 
     test('Åpner vurdert periode - god tro - beløp i behold', async () => {
-        const { getByLabelText, getByText } = renderVilkårsvurderingPeriodeSkjema(
-            lagVilkårsvurderingPeriodeSkjemaData({
-                begrunnelse: 'Gitt i god tro',
-                vilkårsvurderingsresultatInfo: {
-                    vilkårsvurderingsresultat: Vilkårsresultat.GodTro,
-                    godTro: {
-                        begrunnelse: 'Deler av beløpet er i behold',
-                        beløpErIBehold: true,
-                        beløpTilbakekreves: 699,
-                    },
+        const periode = lagVilkårsvurderingPeriodeSkjemaData({
+            begrunnelse: 'Gitt i god tro',
+            vilkårsvurderingsresultatInfo: {
+                vilkårsvurderingsresultat: Vilkårsresultat.GodTro,
+                godTro: {
+                    begrunnelse: 'Deler av beløpet er i behold',
+                    beløpErIBehold: true,
+                    beløpTilbakekreves: 699,
                 },
-            }),
-            true
-        );
-
-        expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+            },
+        });
+        const { getByLabelText } = renderVilkårsvurderingPeriodeSkjema({
+            periode: periode,
+            erTotalbeløpUnder4Rettsgebyr: true,
+        });
+        await waitFor(() => {});
         expect(getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor')).toHaveValue(
             'Gitt i god tro'
         );
@@ -996,35 +1164,33 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
     });
 
     test('Åpner vurdert periode - mangelfulle - uaktsomt - under 4 rettsgebyr', async () => {
-        const { getByLabelText, getByTestId, getByText } = renderVilkårsvurderingPeriodeSkjema(
-            lagVilkårsvurderingPeriodeSkjemaData({
-                begrunnelse: 'Gitt mangelfulle opplysninger',
-                vilkårsvurderingsresultatInfo: {
-                    vilkårsvurderingsresultat: Vilkårsresultat.MangelfulleOpplysningerFraBruker,
-                    aktsomhet: {
-                        begrunnelse: 'Vurdert aktsomhet til uaktsomt',
-                        aktsomhet: Aktsomhet.Uaktsomt,
-                        unnlates4Rettsgebyr: SkalUnnlates.Tilbakekreves,
-                        særligeGrunnerBegrunnelse: 'Det finnes særlige grunner',
-                        særligeGrunner: [
-                            { særligGrunn: SærligeGrunner.GradAvUaktsomhet },
-                            { særligGrunn: SærligeGrunner.StørrelseBeløp },
-                            {
-                                særligGrunn: SærligeGrunner.Annet,
-                                begrunnelse: 'Dette er en annen begrunnelse',
-                            },
-                        ],
-                        særligeGrunnerTilReduksjon: true,
-                        andelTilbakekreves: 33,
-                    },
+        const periode = lagVilkårsvurderingPeriodeSkjemaData({
+            begrunnelse: 'Gitt mangelfulle opplysninger',
+            vilkårsvurderingsresultatInfo: {
+                vilkårsvurderingsresultat: Vilkårsresultat.MangelfulleOpplysningerFraBruker,
+                aktsomhet: {
+                    begrunnelse: 'Vurdert aktsomhet til uaktsomt',
+                    aktsomhet: Aktsomhet.Uaktsomt,
+                    unnlates4Rettsgebyr: SkalUnnlates.Tilbakekreves,
+                    særligeGrunnerBegrunnelse: 'Det finnes særlige grunner',
+                    særligeGrunner: [
+                        { særligGrunn: SærligeGrunner.GradAvUaktsomhet },
+                        { særligGrunn: SærligeGrunner.StørrelseBeløp },
+                        {
+                            særligGrunn: SærligeGrunner.Annet,
+                            begrunnelse: 'Dette er en annen begrunnelse',
+                        },
+                    ],
+                    særligeGrunnerTilReduksjon: true,
+                    andelTilbakekreves: 33,
                 },
-            }),
-            true
-        );
-
-        await waitFor(() => {
-            expect(getByText('Detaljer for valgt periode')).toBeInTheDocument();
+            },
         });
+        const { getByLabelText, getByTestId, getByText } = renderVilkårsvurderingPeriodeSkjema({
+            periode: periode,
+            erTotalbeløpUnder4Rettsgebyr: true,
+        });
+        await waitFor(() => {});
         expect(getByLabelText('Begrunn hvorfor du valgte vilkåret ovenfor')).toHaveValue(
             'Gitt mangelfulle opplysninger'
         );
@@ -1060,20 +1226,81 @@ describe('VilkårsvurderingPeriodeSkjema', () => {
         expect(getByTestId('andelSomTilbakekrevesManuell')).toHaveValue('33');
     });
 
-    test('Validering vises når man forsøker å gå videre uten å fylle inn påkrevde felter', async () => {
-        const { getByRole, queryAllByText } = renderVilkårsvurderingPeriodeSkjema(
-            lagVilkårsvurderingPeriodeSkjemaData(),
-            false
-        );
+    test('Viser særlige grunner og for over 4 rettsgebyr alternativ - uaktsomt', async () => {
+        const periode = lagVilkårsvurderingPeriodeSkjemaData({
+            begrunnelse: 'Gitt mangelfulle opplysninger',
+            vilkårsvurderingsresultatInfo: {
+                vilkårsvurderingsresultat: Vilkårsresultat.MangelfulleOpplysningerFraBruker,
+                aktsomhet: {
+                    begrunnelse: 'Vurdert aktsomhet til uaktsomt',
+                    aktsomhet: Aktsomhet.Uaktsomt,
+                    unnlates4Rettsgebyr: SkalUnnlates.Over4Rettsgebyr,
+                    særligeGrunnerBegrunnelse: 'Det finnes særlige grunner',
+                    særligeGrunner: [
+                        { særligGrunn: SærligeGrunner.GradAvUaktsomhet },
+                        { særligGrunn: SærligeGrunner.StørrelseBeløp },
+                        {
+                            særligGrunn: SærligeGrunner.Annet,
+                            begrunnelse: 'Dette er en annen begrunnelse',
+                        },
+                    ],
+                    særligeGrunnerTilReduksjon: true,
+                    andelTilbakekreves: 33,
+                },
+            },
+        });
+        const { getByTestId, getByText } = renderVilkårsvurderingPeriodeSkjema({
+            periode: periode,
+            behandling: lagBehandling({ erNyModell: true }),
+            erTotalbeløpUnder4Rettsgebyr: true,
+        });
+        expect(
+            getByText(
+                'Totalbeløpet kan være under 4 ganger rettsgebyret (6. ledd). Skal det tilbakekreves?'
+            )
+        ).toBeInTheDocument();
+        expect(
+            getByTestId('tilbakekrevSelvOmBeloepErUnder4Rettsgebyr_Over4Rettsgebyr')
+        ).toBeChecked();
+        expect(
+            getByText('Hvilke særlige grunner kan være aktuelle i denne saken?')
+        ).toBeInTheDocument();
+    });
 
-        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(0);
-
-        await user.click(
-            getByRole('button', {
-                name: 'Gå videre til vedtakssteget',
-            })
-        );
-
-        expect(queryAllByText('Feltet må fylles ut')).toHaveLength(2);
+    test('Viser ikke over 4 rettsgebyr alternativ for gammel modell', async () => {
+        const periode = lagVilkårsvurderingPeriodeSkjemaData({
+            begrunnelse: 'Gitt mangelfulle opplysninger',
+            vilkårsvurderingsresultatInfo: {
+                vilkårsvurderingsresultat: Vilkårsresultat.MangelfulleOpplysningerFraBruker,
+                aktsomhet: {
+                    begrunnelse: 'Vurdert aktsomhet til uaktsomt',
+                    aktsomhet: Aktsomhet.Uaktsomt,
+                    unnlates4Rettsgebyr: SkalUnnlates.Unnlates,
+                    særligeGrunnerBegrunnelse: 'Det finnes særlige grunner',
+                    særligeGrunner: [
+                        { særligGrunn: SærligeGrunner.GradAvUaktsomhet },
+                        { særligGrunn: SærligeGrunner.StørrelseBeløp },
+                        {
+                            særligGrunn: SærligeGrunner.Annet,
+                            begrunnelse: 'Dette er en annen begrunnelse',
+                        },
+                    ],
+                    særligeGrunnerTilReduksjon: true,
+                    andelTilbakekreves: 33,
+                },
+            },
+        });
+        const { queryByTestId, getByText } = renderVilkårsvurderingPeriodeSkjema({
+            periode: periode,
+            erTotalbeløpUnder4Rettsgebyr: true,
+        });
+        expect(
+            getByText(
+                'Totalbeløpet kan være under 4 ganger rettsgebyret (6. ledd). Skal det tilbakekreves?'
+            )
+        ).toBeInTheDocument();
+        expect(
+            queryByTestId('tilbakekrevSelvOmBeloepErUnder4Rettsgebyr_Over4Rettsgebyr')
+        ).not.toBeInTheDocument();
     });
 });
